+++ public/data/cBrewcoffeeList.json
@@ -58,17 +58,5 @@
     "imgAlt":"cb-10.jpg",
     "isIce": true,
     "isLike":false
- },{"id":"BIceCoffee",
-    "name":"아이스 커피",
-    "imgUrl":"../images/b-01.jpg",
-    "imgAlt":"b-01.jpg",
-    "isIce": true,
-    "isLike":false
- },{"id":"BTodayCoffee",
-    "name":"오늘의 커피",
-    "imgUrl":"../images/b-02.jpg",
-    "imgAlt":"b-02.jpg",
-    "isIce": false,
-    "isLike":false
  }
 ]
